Pass --force to `git checkout`
I was experiencing an issue where the repos.checkout() function would
produce a working tree that was missing lots of files. They were not
present on disk, and `git status` marked them as being deleted in the
working tree.

The --force option works around whatever bug was causing this, and there
is no danger of overwriting any local changes with it in this case.