Use `git clone` instead of the copy_repo() function
The only issue with this approach is that the 'origin' remote will
point at the repo cache (file:///src/cache/gits/...) instead of the
Git server (git://git.baserock.org/...). That's not a problem when
we are running a build, as the build shouldn't be able to access the
network anyway.