assembly: Fix build-system handling
Morph treats build-system being missing as being build-system: manual.
YBD treats it as not having a build system.

This was previously acceptable as build-system: manual was a no-op, but
I want to add strip commands to everything by default, and adding them
to the build systems is the appropriate place.

This is because YBD took an absence of build-system or any commands as
being the same as not having a definition file, hence cause to detect
the build system.

This is fine, but the implementation assumed that this meant there was
no build system, hence didn't need to merge the commands.

In the process of fixing this, the check for not having a definition
file was made more explicit, and the system integration logic more
separate.