Split up deploy() to make the recursion less confusing
Also, correctly deploy the subsystems. The subsystems must be deployed
before the parent but after the parent system has been extracted, as
the location field for subsystems is relative to the location of the
parent system. This means that the subsystem needs to know where the
parent system has been extracted to.