Treat True and False as actual commands, rather than booleans.
It's quite easy to forget that "true" and "false" are special words in YAML,
and write "false" to halt the build process while debugging. This was treated
as a Boolean instead of a string, so causes an unhelpful error in ybd.

Since there is no use for booleans in morphologies, they should be treated as
strings.