Adds Riemann support
This patch will optionally cause ybd to emit messages to a Riemann
server configured in the ybd.conf file. It will only execute if the
riemann-client is available on the system.

Two services are created:

  * Timer - this will execute for most "Elapsed time events"
  * Artifact_Timer - this shows the time for an artifact to be created

Timer's description is the text of the trigger that has caused a timer
to be displayed by ybd. Artifact_Timer's description will be the name
of the given repo.