cache.py: Use shutil.move() instead of os.rename()
This will fix cases where the configured artifacts directory
is not on the same file system as the tmp directory where
builds are performed.

Some of the changes in this file are not required, for instance
when we rename the tar file in the same directory, however for
consistency it's better that we use shutil.move() everywhere
in one unilateral change.