utils.py: Dont ever set the timestamp of symlink targets.
If we are staging the target, we will also set it's timestamp,
if the target is invalid, it would fail (hence the previous, removed
check), if the target is valid, it likely points to somewhere on
your system, like /usr/share/aclocal/gtk-doc.m4. Building should
not cause your host system files to be modified.

This was previously going unnoticed due to builds being run
as root.