Fix ybd representation of splits
Previously ybd's definitions model had accidentally left
foo['strata'] and foo['chunks'] lying around, and splitting used
that data. This made things very confusing.

Now we pop that content into the applicable child definitions,
and as a result need to explicitly keep track of the splits in the
parent definition.