Ensure that we only stage relative paths for symbolic links.
If we have artifacts which contain symbolic links to absolute paths,
staging any files under those paths will attempt to write to the
build host filesystem instead of the staging directory.

Work around this by simply ensuring that all symbolic links created
in the staging directory are created with relative paths, so that
they resolve to the correct location whether or not we are in an
active chroot / sandbox.