utils.py: Ensure realpath exists for staged broken symlinks.
This is important because a chunk can install a link to a
directory that does not exist; while not being responsible
for creating the symlink target directory. The result is
later failures when trying to stage files across the broken
symlink boundary.

This throws an IOError (and not a clean exit()) if a symlink
points outside of the staging area. This is because resolving
the relative symlink path is taken care of so if the error
triggers it's an internal bug (not a problem with definitions).