utils.py: Warning when source file listed in artifact does not exist
We use the split artifact metadata to move files from the staged
system directory into a <system>.inst/ subdirectory one file at
a time, after running system integration commands and ldconfig.

The current working assumption is that if a file was listed in
the artifact output; it will certainly exist when we try to move
it into the system install directory, but this is not true.

Running system integration commands and ldconfig can result in
removing files, one working example is when you have broken symlink
installed into an artifact inside a libdir. ldconfig will delete
dangling symlinks in it's libdirs with extreme prejudice.

It may or may not make sense to eventually turn this into an error.